Store theme tints as f64 fractions and keep theme colors through reads
The i8 percentage quantization corrupted Excel tints like
0.3499862666707358 on every round trip, and DrawingML lumMod/lumOff
values lost their sub-percent precision. Tint is now the OOXML f64
fraction end to end; Color keeps total equality by comparing tints
bitwise.

Readers also stop baking theme colors into RGB at read time: styles
and conditional formats keep Color::Theme, the palette stays on the
workbook, and display RGB comes from Workbook::resolve_color.

+++ bindings/nodejs/src/drawings.rs
@@ -290,7 +290,7 @@ enum DrawingColor {
     Auto,
     Rgb { r: u8, g: u8, b: u8 },
     Argb { a: u8, r: u8, g: u8, b: u8 },
-    Theme { index: u8, tint: i8 },
+    Theme { index: u8, tint: f64 },
     Indexed { index: u8 },
 }
