+++ bindings/wasm/src/drawings.rs
@@ -321,7 +321,7 @@ pub(crate) enum WasmDrawingColor {
     Auto,
     Rgb { r: u8, g: u8, b: u8 },
     Argb { a: u8, r: u8, g: u8, b: u8 },
-    Theme { index: u8, tint: i8 },
+    Theme { index: u8, tint: f64 },
     Indexed { index: u8 },
 }
